// src/app/funcionario/model/funcionario.ts
export interface Funcionario {
    
    id: string,
    nomeFun: string;  
    cpf: string;    
    rg: string;    
    telefone?: string;    
    sexoFun?: string;    
    dataNasciFun: string;    
    tituloEleitor?: string;    
    estadoCivil?: string;
    grauInstrucao?: string;
    rua?: string;
    bairro?: string;
    cep?: string;
    numero?: string;
    city?: string;
    funcao: string;
    salario: number;
    dataAdmissao: string;
    horarioTrabalho?: string;
    intervaloTrabalho?: string;
    contratoExpe?: number;
    valeTrans?: number;
    valeAlimen?: number;
    numCarteiraTrab?: string;
    serieCarteiraTrab?: string;
    estadoCarteiraTrab?: string;
}

export interface Funcionario2 {
    
    nomeFun: string;  
    cpf: string;    
    rg: string;    
    telefone?: string;    
    sexoFun?: string;    
    dataNasciFun: string;    
    tituloEleitor?: string;    
    estadoCivil?: string;
    grauInstrucao?: string;
    rua?: string;
    bairro?: string;
    cep?: string;
    numero?: string;
    city?: string;
    funcao: string;
    salario: number;
    dataAdmissao: string;
    horarioTrabalho?: string;
    intervaloTrabalho?: string;
    contratoExpe?: number;
    valeTrans?: number;
    valeAlimen?: number;
    numCarteiraTrab?: string;
    serieCarteiraTrab?: string;
    estadoCarteiraTrab?: string;
}